refactored project heirarchy
