+++ src/Graph.DB/GraphType.cs
