+++ src/Graph.DB/MemoryGraph.cs
