+++ src/Graph.Indexes/AdjacencyList.cs
