+++ src/Graph.Indexes/AdjacencyMatrix.cs
