+++ src/Graph.Indexes/BinaryAdjacencyMatrix.cs
