+++ src/Graph.Indexes/DirectedAdjacencyMatrix.cs
@@ -1,5 +1,4 @@
-﻿using Graph.Graphs;
-using System;
+﻿using System;
 
 namespace Graph.Indexes
 {
@@ -60,6 +59,6 @@ namespace Graph.Indexes
             return false;
         }
 
-        public override GraphType Type => GraphType.Directed;
+        public override IndexType Type => IndexType.Directed;
     }
 }
