+++ src/Graph.Indexes/IndexType.cs
@@ -1,0 +1,8 @@
+﻿namespace Graph.Indexes
+{
+    public enum IndexType
+    {
+        Directed,
+        Undirected,
+    }
+}
