+++ src/Graph.Indexes/UndirectedAdjacencyList.cs
@@ -1,5 +1,4 @@
-﻿using Graph.Graphs;
-using System;
+﻿using System;
 using System.Collections.Generic;
 
 namespace Graph.Indexes
@@ -64,6 +63,6 @@ namespace Graph.Indexes
             return false;
         }
 
-        public override GraphType Type => GraphType.Undirected;
+        public override IndexType Type => IndexType.Undirected;
     }
 }
