+++ src/Graph.Traversals/DepthFirstPreOrderTraversal.cs
@@ -5,24 +5,22 @@ using System.Linq;
 
 namespace Graph.Traversals
 {
-    public sealed class DepthFirstPostOrderTraversal<TKey>
+    public sealed class DepthFirstPreOrderTraversal<TKey>
         : Traversal<TKey>
-    where TKey : IComparable, IComparable<TKey>, IEquatable<TKey>
+        where TKey : IComparable, IComparable<TKey>, IEquatable<TKey>
     {
-        public DepthFirstPostOrderTraversal(IAdjacencyIndex<TKey> adjacencyIndex)
+        public DepthFirstPreOrderTraversal(IAdjacencyIndex<TKey> adjacencyIndex)
             : base(adjacencyIndex)
         {
         }
 
         public override IEnumerable<TKey> Traverse(TKey vertex)
         {
-            // todo: MSL - make sure I don't have to call yield return to allow for the yields in the call to traverse
             return this.Traverse(vertex, -1);
         }
 
         public override IEnumerable<TKey> Traverse(TKey vertex, int depth)
         {
-            var visitStack = new Stack<TKey>();
             var visited = new HashSet<TKey>(this.AdjacencyIndex.Size);
             var neighbors = new Stack<TKey>(new TKey[] { vertex });
 
@@ -31,18 +29,13 @@ namespace Graph.Traversals
                 var nextVertex = neighbors.Pop();
                 if (!visited.Contains(nextVertex))
                 {
-                    visitStack.Push(nextVertex);
+                    yield return nextVertex;
                     visited.Add(nextVertex);
                     foreach (var neighbor in this.AdjacencyIndex.Neighbors(nextVertex).Where(n => !visited.Contains(n)))
                     {
                         neighbors.Push(neighbor);
                     }
                 }
-            }
-
-            while (visitStack.Count > 0)
-            {
-                yield return visitStack.Pop();
             }
         }
     }
