+++ src/Graph.Traversals/FastDepthFirstPostOrderTraversal.cs
