+++ src/Graph.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31515.178
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Graph.Lib", "graph.lib\Graph.Lib.csproj", "{5FD338E4-8D5C-4F6D-B802-A93F4729C70C}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{713B991B-9264-447C-B5E5-A31D73B99A9C}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
@@ -15,16 +13,20 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Graph.Test", "Graph.Test\Gr
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Graph.DB", "Graph.DB\Graph.DB.csproj", "{5456F1BB-02C2-414C-B5CA-8EBBAB778AC2}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Graph.Indexes", "Graph.Indexes\Graph.Indexes.csproj", "{374E6536-BDAD-4AA5-9563-32FA45C8710D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Graph.Traversals", "Graph.Traversals\Graph.Traversals.csproj", "{4A7F6535-3700-4542-BF91-7716340196DF}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{9915047B-845D-4E77-B80C-DCDA75268A80}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{4BF8A6DB-C6FF-40BA-A190-39EC0D8E1AE4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5FD338E4-8D5C-4F6D-B802-A93F4729C70C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5FD338E4-8D5C-4F6D-B802-A93F4729C70C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5FD338E4-8D5C-4F6D-B802-A93F4729C70C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5FD338E4-8D5C-4F6D-B802-A93F4729C70C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5FF66DFA-8539-40F8-AFBA-626BC8E8227A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5FF66DFA-8539-40F8-AFBA-626BC8E8227A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5FF66DFA-8539-40F8-AFBA-626BC8E8227A}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -33,9 +35,23 @@ Global
 		{5456F1BB-02C2-414C-B5CA-8EBBAB778AC2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5456F1BB-02C2-414C-B5CA-8EBBAB778AC2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5456F1BB-02C2-414C-B5CA-8EBBAB778AC2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{374E6536-BDAD-4AA5-9563-32FA45C8710D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{374E6536-BDAD-4AA5-9563-32FA45C8710D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{374E6536-BDAD-4AA5-9563-32FA45C8710D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{374E6536-BDAD-4AA5-9563-32FA45C8710D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4A7F6535-3700-4542-BF91-7716340196DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4A7F6535-3700-4542-BF91-7716340196DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4A7F6535-3700-4542-BF91-7716340196DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4A7F6535-3700-4542-BF91-7716340196DF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{5FF66DFA-8539-40F8-AFBA-626BC8E8227A} = {4BF8A6DB-C6FF-40BA-A190-39EC0D8E1AE4}
+		{5456F1BB-02C2-414C-B5CA-8EBBAB778AC2} = {9915047B-845D-4E77-B80C-DCDA75268A80}
+		{374E6536-BDAD-4AA5-9563-32FA45C8710D} = {9915047B-845D-4E77-B80C-DCDA75268A80}
+		{4A7F6535-3700-4542-BF91-7716340196DF} = {9915047B-845D-4E77-B80C-DCDA75268A80}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D4FF5415-98E3-4C9E-A10D-75A2795F81C3}
